split halo model into 1/2/4-pt
