+++ pyccl/halos/pk_1pt.py
@@ -1,0 +1,111 @@
+from .profiles import HaloProfile
+import numpy as np
+
+
+__all__ = ("halomod_mean_profile_1pt", "halomod_bias_1pt",)
+
+
+def halomod_mean_profile_1pt(cosmo, hmc, k, a, prof,
+                             normprof=False):
+    """ Returns the mass-weighted mean halo profile.
+
+    .. math::
+        I^0_1(k,a|u) = \\int dM\\,n(M,a)\\,\\langle u(k,a|M)\\rangle,
+
+    where :math:`n(M,a)` is the halo mass function, and
+    :math:`\\langle u(k,a|M)\\rangle` is the halo profile as a
+    function of scale, scale factor and halo mass.
+
+    Args:
+        cosmo (:class:`~pyccl.core.Cosmology`): a Cosmology object.
+        hmc (:class:`HMCalculator`): a halo model calculator.
+        k (float or array_like): comoving wavenumber in Mpc^-1.
+        a (float or array_like): scale factor.
+        prof (:class:`~pyccl.halos.profiles.HaloProfile`): halo
+            profile.
+        normprof (bool): if `True`, this integral will be
+            normalized by :math:`I^0_1(k\\rightarrow 0,a|u)`.
+
+    Returns:
+        float or array_like: integral values evaluated at each
+        combination of `k` and `a`. The shape of the output will
+        be `(N_a, N_k)` where `N_k` and `N_a` are the sizes of
+        `k` and `a` respectively. If `k` or `a` are scalars, the
+        corresponding dimension will be squeezed out on output.
+    """
+    a_use = np.atleast_1d(a).astype(float)
+    k_use = np.atleast_1d(k).astype(float)
+
+    # Check inputs
+    if not isinstance(prof, HaloProfile):
+        raise TypeError("prof must be of type `HaloProfile`")
+
+    na = len(a_use)
+    nk = len(k_use)
+    out = np.zeros([na, nk])
+    for ia, aa in enumerate(a_use):
+        i01 = hmc.I_0_1(cosmo, k_use, aa, prof)
+        if normprof:
+            norm = hmc.profile_norm(cosmo, aa, prof)
+            i01 *= norm
+        out[ia, :] = i01
+
+    if np.ndim(a) == 0:
+        out = np.squeeze(out, axis=0)
+    if np.ndim(k) == 0:
+        out = np.squeeze(out, axis=-1)
+    return out
+
+
+def halomod_bias_1pt(cosmo, hmc, k, a, prof, normprof=False):
+    """ Returns the mass-and-bias-weighted mean halo profile.
+
+    .. math::
+        I^1_1(k,a|u) = \\int dM\\,n(M,a)\\,b(M,a)\\,
+        \\langle u(k,a|M)\\rangle,
+
+    where :math:`n(M,a)` is the halo mass function,
+    :math:`b(M,a)` is the halo bias, and
+    :math:`\\langle u(k,a|M)\\rangle` is the halo profile as a
+    function of scale, scale factor and halo mass.
+
+    Args:
+        cosmo (:class:`~pyccl.core.Cosmology`): a Cosmology object.
+        hmc (:class:`HMCalculator`): a halo model calculator.
+        k (float or array_like): comoving wavenumber in Mpc^-1.
+        a (float or array_like): scale factor.
+        prof (:class:`~pyccl.halos.profiles.HaloProfile`): halo
+            profile.
+        normprof (bool): if `True`, this integral will be
+            normalized by :math:`I^0_1(k\\rightarrow 0,a|u)`
+            (see :meth:`~HMCalculator.I_0_1`).
+
+    Returns:
+        float or array_like: integral values evaluated at each
+        combination of `k` and `a`. The shape of the output will
+        be `(N_a, N_k)` where `N_k` and `N_a` are the sizes of
+        `k` and `a` respectively. If `k` or `a` are scalars, the
+        corresponding dimension will be squeezed out on output.
+    """
+    a_use = np.atleast_1d(a).astype(float)
+    k_use = np.atleast_1d(k).astype(float)
+
+    # Check inputs
+    if not isinstance(prof, HaloProfile):
+        raise TypeError("prof must be of type `HaloProfile`")
+
+    na = len(a_use)
+    nk = len(k_use)
+    out = np.zeros([na, nk])
+    for ia, aa in enumerate(a_use):
+        i11 = hmc.I_1_1(cosmo, k_use, aa, prof)
+        if normprof:
+            norm = hmc.profile_norm(cosmo, aa, prof)
+            i11 *= norm
+        out[ia, :] = i11
+
+    if np.ndim(a) == 0:
+        out = np.squeeze(out, axis=0)
+    if np.ndim(k) == 0:
+        out = np.squeeze(out, axis=-1)
+    return out
